requirements.txt
numpy==1.16.1
PyQt5==5.11.3
PyQt5-sip==4.19.13
pyqtgraph==0.10.0
PyVISA==1.9.1
PyVISA-py==0.3.1
PyVISA-sim==0.3
PyYAML==3.13
scipy==1.2.0
stringparser==0.4.1
numba
